Update and rename OrbitalTeleOp.java to OrbitalTeleOpv1.java
Deleted some garbage + version control
